*: add support for [sparse] fully heirarchical partitions
This commit extends the original PR
https://github.com/rohany/taco/pull/47 to allow for expression of
partitioning a sparse tensor ahead of time. It additionally reworks some
awkwardness and bugs around the original implementation of up front
partitioning. In particular, the old up front partitioning approach
would require there to be a partition for each of the tensors, which
isn't necessarily true if a computation ends up replicating a tensor.