petsc: switch to a hand-installed version of petsc
This switch was necessary because the Lassen-provided version of PETSc
was not compiled with optimization flags. This resulted in a significant
performance degradation when compared to DISTAL and TACO. This commit adds
PETSc as a submodule and includes configuration and installation information
for PETSc. We try to follow the Lassen configuration as closely as possible,
except for adding optimization flags to the configuration.