*: support GPU workspaces local to an FB memory in DISTAL
This commit includes several bugfixes and tying up of loose ends to
support workspaces in DISTAL (which have been largely ignored by me
since the beginning of the project). The commits have been tested only
for GPUs, so it's likely that the current code does not work for CPUs,
but my experience so far has indicated that the bugs around this area
are small, so if some kernels need a CPU workspace then further work can
be put in to fix those bugs.